EY_backend/ey-flask/requirements.txt
Flask
Flask-CORS
python-dotenv
google-generativeai
gunicorn
